CRUD of food item done

+++ src/main/java/cs544/fooddelivery/domain/Category.java
@@ -10,20 +10,19 @@ import javax.persistence.OneToMany;
 @Entity
 public class Category {
 	
-	@Id
-	@GeneratedValue
-	private int id;
+	@Id @GeneratedValue
+	private Long id;
 	
 	private String name;
 	
 	@OneToMany(mappedBy="category")
 	private List<FoodItem> foodItems;
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
